remove define, compile rules immediately
grit.define removed so no later rules can be defined. Name changed to _define

_compile is called immediately after the rules are defined.

Thus grammar errors are reported when the rules are defined, as against when the first .parse() is called.

semantic action functions can not be resolved until the first .prase() so there may be a run time error if any semantic actions refer to functions that remain undefined.